refs #699: move admin listing into course management
