+++ .beads/interactions.jsonl
@@ -22,3 +22,4 @@
 {"id":"int-976060c78e31c3b0c10920bb9986c375","kind":"field_change","created_at":"2026-07-05T19:14:17.134169Z","actor":"Simson Garfinkel","issue_id":"webapp-cgr.20","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Implemented make course-create and dbutil create-course with interactive prompts, existing-admin selection, app.course_management reusable course/admin workflow, concise noninteractive usage, idempotent retry behavior, docs, and focused tests. Also addressed Copilot review edge cases for nullable course names and JSON-safe template globals before commit."}}
 {"id":"int-23e819c9366e5d116565633000b04847","kind":"field_change","created_at":"2026-07-05T19:20:37.377878Z","actor":"Simson Garfinkel","issue_id":"webapp-cgr.21","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Implemented make demo-course-create via dbutil create-demo-course, moved reusable demo-course initialization into app.course_management, added idempotency/parser tests, and documented local/deployed demo usage."}}
 {"id":"int-e5a3923113e4688fd4dba80687ac4f2d","kind":"field_change","created_at":"2026-07-05T19:24:19.686866Z","actor":"Simson Garfinkel","issue_id":"webapp-cgr.16","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Implemented lambda-only post-deploy course initialization as make sam-course-create, deriving DynamoDBTablePrefix/ApplicationUrl/MailerDryRun from the selected SAM stack, and documented the replacement for VM bootstrap section 10 in the cutover runbook."}}
+{"id":"int-f2528b24bd1c127764877e24d4bc2415","kind":"field_change","created_at":"2026-07-05T20:10:52.284846Z","actor":"Simson Garfinkel","issue_id":"webapp-cgr.22","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Moved course administrator read model into app.course_management; dbutil now consumes service-layer list_admins; focused tests, docs build, and pylint pass."}}
