+++ src/app/odb.py
@@ -15,7 +15,6 @@ import functools
 import hashlib
 import uuid
 import time
-from collections import defaultdict
 from functools import wraps
 from decimal import Decimal
 
@@ -29,8 +28,6 @@ from .schema import (
     Movie,
     Trackpoint,
     RenameMarkerRequest,
-    AdminCourse,
-    CourseAdmin,
     validate_movie_field,
     Course,
     fix_movie,
@@ -1012,44 +1009,6 @@ def list_users_courses(*, user_id):
 
     return {USERS: users_list, COURSES: courses_list}
 
-
-def list_admins():
-    """Return all course administrators and the courses they administer."""
-    dd = DDBO()
-    courses_by_admin: dict[str, list[AdminCourse]] = defaultdict(list)
-    last_evaluated_key = None
-
-    while True:
-        scan_kwargs = {}
-        if last_evaluated_key:
-            scan_kwargs['ExclusiveStartKey'] = last_evaluated_key
-
-        response = dd.courses.scan(**scan_kwargs)
-        for course in response.get('Items', []):
-            admin_course = AdminCourse(
-                course_id=course[COURSE_ID],
-                course_name=course.get(COURSE_NAME) or "",
-            )
-            for user_id in course.get(ADMINS_FOR_COURSE, []):
-                courses_by_admin[user_id].append(admin_course)
-        last_evaluated_key = response.get('LastEvaluatedKey')
-        if not last_evaluated_key:
-            break
-
-    admins = []
-    for user_id, courses in courses_by_admin.items():
-        try:
-            user = dd.get_user(user_id)
-        except InvalidUser_Id:
-            logger.warning("course admin user_id %s is missing from users table", user_id)
-            continue
-        admins.append(CourseAdmin(
-            user_id=user_id,
-            email=user.get(EMAIL, ""),
-            user_name=user.get(USER_NAME, ""),
-            courses=sorted(courses, key=lambda course: (course.course_name, course.course_id)),
-        ))
-    return sorted(admins, key=lambda admin: (admin.user_name.casefold(), admin.email.casefold(), admin.user_id))
 
 # pylint-x: disable=too-many-arguments
 def register_email(email, user_name, *, course_key=None, course_id=None, admin=False):
